+++ src/Application/Input.cpp
@@ -174,13 +174,13 @@ void InputBackend::onEvent(uint8_t* keys, uint32_t numKeys,
 
                 //Update only if needed.
             if (uint32_t(newWidth) != Window::instance()->width || uint32_t(newHeight) != Window::instance()->height)
-                {
-                    Window::instance()->resizeRequested = true;
-                Window::instance()->width = uint32_t(newWidth);
-                Window::instance()->height = uint32_t(newHeight);
+            {
+                Window::instance()->resizeRequested = true;
+                Window::instance()->width = uint16_t(newWidth);
+                Window::instance()->height = uint16_t(newHeight);
 
                     vprint("Resize to %u, %u\n", Window::instance()->width, Window::instance()->height);
-                }
+            }
             break;
             }
         case SDL_EVENT_WINDOW_FOCUS_GAINED:
